OPENJPA-203.
git-svn-id: https://svn.apache.org/repos/asf/incubator/openjpa/trunk@525950 13f79535-47bb-0310-9956-ffa450edef68

+++ openjpa-jdbc/src/main/java/org/apache/openjpa/jdbc/kernel/PessimisticLockManager.java
@@ -86,7 +86,7 @@ public class PessimisticLockManager
             setLockLevel(sm, LOCK_DATASTORE_ONLY);
     }
 
-    protected void lockInternal(OpenJPAStateManager sm, int level, int timeout,
+    protected void lockInternal(OpenJPAStateManager sm, int level, long timeout,
         Object sdata) {
         // we can skip any already-locked instance regardless of level because
         // we treat all locks the same (though super doesn't)
@@ -103,7 +103,7 @@ public class PessimisticLockManager
      * Lock the specified instance row by issuing a "SELECT ... FOR UPDATE"
      * statement.
      */
-    private void lockRow(OpenJPAStateManager sm, int timeout) {
+    private void lockRow(OpenJPAStateManager sm, long timeout) {
         // assert that the dictionary supports the "SELECT ... FOR UPDATE"
         // construct; if not, and we the assertion does not throw an
         // exception, then just return without locking
@@ -136,7 +136,7 @@ public class PessimisticLockManager
                     if (log.isWarnEnabled())
                         log.warn(_loc.get("millis-query-timeout"));
                 }
-                stmnt.setQueryTimeout(timeout / 1000);
+                stmnt.setQueryTimeout((int) (timeout / 1000));
             }
             rs = stmnt.executeQuery();
             if (!rs.next())
